+++ plugins/device/adm3A-terminal/src/main/java/net/emustudio/plugins/device/adm3a/interaction/LoadCursorPosition.java
@@ -48,11 +48,11 @@ public class LoadCursorPosition {
      * @param data received char
      * @return true if in bounds
      */
-    private boolean checkBounds(Short data) {
+    private boolean checkBounds(Byte data) {
         return data >= ' ' && data <= 'o';
     }
 
-    boolean notAccepted(Short data) {
+    boolean notAccepted(Byte data) {
         if (expect == ESCAPE && data == ASCII_ESC) {
             expect = ASSIGN;
             return false;
